requirements.txt
amqp==2.2.2
billiard==3.5.0.3
blessings==1.6.1
boto==2.48.0
bpython==0.17
cairocffi==0.8.0
CairoSVG==2.1.3
celery==4.1.0
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.4.2
cssmin==0.2.0
cssselect2==0.2.1
curtsies==0.2.11
defusedxml==0.5.0
Django==1.11.9
django-allauth==0.34.0
django-assets==0.12
django-auditlog==0.4.5
django-compat==1.0.15
django-debug-toolbar==1.9.1
django-environ==0.4.4
django-extra-fields==0.9
django-filter==1.1.0
django-hijack==2.1.6
django-hijack-admin==2.1.6
django-jsonfield==1.0.1
django-model-utils==3.1.1
django-post-office==3.0.4
django-redis==4.8.0
https://github.com/fdemmer/django-weasyprint/archive/v0.4.zip
django-widget-tweaks==1.4.1
djangorestframework==3.7.7
factory-boy==2.9.2
Faker==0.8.10
greenlet==0.4.12
gunicorn==19.7.1
html5lib==1.0.1
idna==2.6
invoke==0.22.0
itypes==1.1.0
Jinja2==2.10
jsonfield==2.0.2
kombu==4.1.0
MarkupSafe==1.0
oauthlib==2.0.6
olefile==0.44
pdfrw==0.4
Pillow==5.0.0
psycopg2==2.7.3.2
pycparser==2.18
Pygments==2.2.0
pyinotify==0.9.6
Pyphen==0.9.4
python-dateutil==2.6.1
python3-openid==3.1.0
pytz==2017.3
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
selenium==3.8.1
six==1.11.0
sqlparse==0.2.4
text-unidecode==1.1
tinycss2==0.6.1
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4
wcwidth==0.1.7
WeasyPrint==0.42
webassets==0.12.1
webencodings==0.5.1
whitenoise==3.3.1
